requirements.txt
crowdstrike_falconpy==1.2.3
gspread==5.7.1
oauth2client==4.1.3
